pyproject.toml
[project]
name = "notion_df"
version = "0.3.0"
description = "Notion-df: schema-rich Notion API wrapper"
authors = [{ name = "dyhn78", email = "dyhn78@gmail.com" }]
requires-python = ">=3.12"

dependencies = [
    "networkx==3.1",
    "beautifulsoup4==4.12.2",
    "python-dateutil==2.8.2",
    "requests<=2.31.0", # 2.32.0 will require urllib3=2
    "html5lib==1.1",
    "urllib3<2", # urllib3~=2 breaks in legacy websites
    "selenium==4.9.1",
    "tenacity==8.2.2",
    "types-requests~=2.31.0",
    "typing_extensions==4.10.0",
    "webdriver-manager==4.0.2",
    "psutil==5.9.6",
    "loguru==0.7.2",
    "pytest==7.3.1",
    "ruff>=0.11.4",
]

[project.optional-dependencies]
test = [
    "mypy==1.3.0",
    "mypy-extensions==1.0.0",
    "pytest==7.3.1",
    "vulture==0.12.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
extend-exclude = [
    "app/action/__routine__.py",
]

[tool.hatch.envs.default]
dependencies = [
    "ruff",
    "pytest",
    "mypy",
    "pylint",
    "vulture",
]

[tool.hatch.envs.default.scripts]
apply = [
    "hatch run apply.ruff",
    "hatch run apply.cloc",
]

test = [
    "hatch run test.vulture",
    "hatch run test.ruff",
    "hatch run test.pylint",
    "hatch run test.mypy",
    "hatch run test.unit",
]

"apply.ruff" = [
    "ruff format notion_df test app",
    "ruff check --fix notion_df test app",
]

"apply.cloc" = "cloc notion_df test app"

"test.vulture" = "vulture"
"test.ruff" = [
    "ruff format --check notion_df test app",
    "ruff check notion_df test app",
]
"test.pylint" = "pylint notion_df/"
"test.mypy" = "mypy notion_df/"
"test.unit" = "pytest tests/"

